+++ src/PRo3D.Viewer/Viewer/ViewerGUI.fs
@@ -125,7 +125,7 @@ module Gui =
                     | Planet.Earth -> "Earth (ellipsoid)"
                     | Planet.JPL   -> "JPL Rover Frame"
                     | Planet.None  -> "None xyz"          
-                    | Planet.XZY   -> "XZY"
+                    | Planet.ENU   -> "ENU"
                     | _ -> "[TextOverlays] missing enum"
                 )